+++ packages/material-composer/src/modules/Acceleration.ts
@@ -1,5 +1,5 @@
 import { Input, Mul, pipe, Pow } from "shader-composer"
-import { ModuleFactory } from "."
+import { ModuleFactory } from ".."
 import { Translate } from "./Translate"
 
 export type AccelerationProps = {
