+++ packages/material-composer/src/modules/Layer.ts
@@ -1,3 +1,3 @@
-import { ModuleFactory } from "."
+import { ModuleFactory } from ".."
 
 export const Layer: ModuleFactory = ({}) => (state) => state
